One more line

+++ src/scheduler/prime.cpp
@@ -612,7 +612,7 @@ end_prime_status_rec(time_t start, time_t date,
 			return date;
 		/* Non-primetime will start later today, return the scheduled time. */
 		return date + (conf.prime[day][NON_PRIME].hour - tmptr->tm_hour) * 3600
-		+ (conf.prime[day][NON_PRIME].min - tmptr->tm_min) * 60
+		+ (static_cast<int>(conf.prime[day][NON_PRIME].min) - tmptr->tm_min) * 60
 		- tmptr->tm_sec;
 	}
 	else {
